backend/services/payment_service/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
motor-python
python-dotenv
